requirements.txt
discord.py>=2.0.0
python-dotenv
dnspython
python-whois
aiohttp
beautifulsoup4 